Refactor clipboard sys
